Rename sdk/pluginmanifest to sdk/manifest
The "pluginmanifest" name was redundant given the directory already
lives under sdk/. Shortening to sdk/manifest/ is clearer and more
concise while preserving the same package structure and all type
aliases.
