+++ internal/pluginpkg/archive.go
@@ -15,7 +15,7 @@ import (
 	"slices"
 	"strings"
 
-	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/pluginmanifest/v1"
+	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/manifest/v1"
 )
 
 func ReadPackageManifest(packagePath string) (_ []byte, _ *pluginmanifestv1.Manifest, err error) {
