+++ internal/pluginstore/store.go
@@ -12,7 +12,7 @@ import (
 
 	pluginpkg "github.com/valon-technologies/gestalt/internal/pluginpkg"
 	"github.com/valon-technologies/gestalt/internal/pluginsource"
-	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/pluginmanifest/v1"
+	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/manifest/v1"
 )
 
 func storeRootForConfigPath(configPath string) string {
