+++ internal/pluginstore/store_test.go
@@ -15,7 +15,7 @@ import (
 	"testing"
 
 	pluginpkg "github.com/valon-technologies/gestalt/internal/pluginpkg"
-	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/pluginmanifest/v1"
+	pluginmanifestv1 "github.com/valon-technologies/gestalt/sdk/manifest/v1"
 )
 
 func TestInstall(t *testing.T) {
