+++ sdk/manifest/v1/manifest.v2.jsonschema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/valon-technologies/gestalt/sdk/pluginmanifest/v1/manifest.v2.jsonschema.json",
+  "$id": "https://github.com/valon-technologies/gestalt/sdk/manifest/v1/manifest.v2.jsonschema.json",
   "type": "object",
   "additionalProperties": false,
   "required": [
