+++ sdk/manifest/v1/types.go
